Stop counting rewards as part of (fully) protected stake
Why make this change?
* Seen a lot of people confused on discord about the
value of their fully protected stake and protected
amount. It is unclear why rewards are counted towards
those.
* The tooltips on the table header do not mention rewards.
* When trying to unstake, the amount you are able to
unstake is lower than the protected amount shown in
the table because that value includes rewards.
* When you expand a collapsed position (multiple stakes
in the same pool) the sum of all individual positions
do not add up to the top one.

This commit fixes all of the above.