Split up ROI on rewards from regular fee based ROI
Why this change?
* If you have multiple positions in the same pool,
the roi shown at the top, should be a weighted sum
of the roi of the individual positions.
* If you at some point claim you rewards, the roi
on your position all of the sudden drops which is
confusing.
* Fees and Rewards are already split, why not split
the roi as well?

By splitting up the roi in the base roi on fees
and one extra roi based on rewards, the sum of
roi's on a expanded position (multiple possitions
in same pool) now holds in the regular part.
And this part does not drop in value once you
claim your rewards.

Splitting of the roi coming from rewards imo is
a nice balance between showing of the protocols
great yield and providing clarity of where that
yield comes from.